Extract library managers
