+++ mainappsrc/core/automl_core.py
@@ -76,11 +76,6 @@ from analysis.mechanisms import (
 )
 from pathlib import Path
 from collections.abc import Mapping
-import csv
-try:
-    from openpyxl import load_workbook
-except Exception:  # openpyxl may not be installed
-    load_workbook = None
 from gui.utils.drawing_helper import FTADrawingHelper, fta_drawing_helper
 from mainappsrc.core.event_dispatcher import EventDispatcher
 from mainappsrc.core.window_controllers import WindowControllers
@@ -2148,422 +2143,8 @@ class AutoMLApp(
     def manage_scenario_libraries(self):
         return self.scenario_library_manager.manage_scenario_libraries()
 
-        def edit_lib():
-            sel = lib_lb.curselection()
-            if not sel:
-                return
-            lib = self.scenario_libraries[sel[0]]
-            dlg = LibraryDialog(win, self, lib)
-            lib.update(dlg.data)
-            refresh_libs()
-
-        def delete_lib():
-            sel = lib_lb.curselection()
-            if sel:
-                idx = sel[0]
-                del self.scenario_libraries[idx]
-                refresh_libs()
-
-        def add_scen():
-            sel = lib_lb.curselection()
-            if not sel:
-                return
-            lib = self.scenario_libraries[sel[0]]
-            dlg = ScenarioDialog(win, self, lib)
-            if dlg.data.get("name"):
-                lib.setdefault("scenarios", []).append(dlg.data)
-                refresh_scenarios()
-
-        def edit_scen():
-            sel_lib = lib_lb.curselection()
-            sel_sc = scen_tree.selection()
-            if not sel_lib or not sel_sc:
-                return
-            lib = self.scenario_libraries[sel_lib[0]]
-            idx = scen_tree.index(sel_sc[0])
-            data = lib.get("scenarios", [])[idx]
-            dlg = ScenarioDialog(win, self, lib, data)
-            lib["scenarios"][idx] = dlg.data
-            refresh_scenarios()
-
-        def del_scen():
-            sel_lib = lib_lb.curselection()
-            sel_sc = scen_tree.selection()
-            if not sel_lib or not sel_sc:
-                return
-            lib = self.scenario_libraries[sel_lib[0]]
-            idx = scen_tree.index(sel_sc[0])
-            del lib.get("scenarios", [])[idx]
-            refresh_scenarios()
-
-        btnf = ttk.Frame(win)
-        btnf.grid(row=1, column=1, columnspan=3, sticky="ew")
-        ttk.Button(btnf, text="Add Lib", command=add_lib).pack(side=tk.LEFT)
-        ttk.Button(btnf, text="Edit Lib", command=edit_lib).pack(side=tk.LEFT)
-        ttk.Button(btnf, text="Del Lib", command=delete_lib).pack(side=tk.LEFT)
-        ttk.Button(btnf, text="Add Scen", command=add_scen).pack(side=tk.LEFT, padx=5)
-        ttk.Button(btnf, text="Edit Scen", command=edit_scen).pack(side=tk.LEFT)
-        ttk.Button(btnf, text="Del Scen", command=del_scen).pack(side=tk.LEFT)
-
-        lib_lb.bind("<<ListboxSelect>>", refresh_scenarios)
-        refresh_libs()
-
     def manage_odd_libraries(self):
-        if hasattr(self, "_odd_tab") and self._odd_tab.winfo_exists():
-            self.doc_nb.select(self._odd_tab)
-            return
-        self._odd_tab = self.lifecycle_ui._new_tab("ODD Libraries")
-        win = self._odd_tab
-        lib_lb = tk.Listbox(win, height=8, width=25)
-        lib_lb.grid(row=0, column=0, rowspan=4, sticky="nsew")
-        elem_tree = ttk.Treeview(win, columns=("cls", "attrs"), show="tree headings")
-        elem_tree.heading("#0", text="Name")
-        elem_tree.column("#0", width=150)
-        elem_tree.heading("cls", text="Class")
-        elem_tree.column("cls", width=120)
-        elem_tree.heading("attrs", text="Attributes")
-        elem_tree.column("attrs", width=200)
-        elem_tree.grid(row=0, column=1, columnspan=3, sticky="nsew")
-        win.grid_rowconfigure(0, weight=1)
-        win.grid_columnconfigure(1, weight=1)
-
-        if not hasattr(self, "odd_elem_icon"):
-            self.odd_elem_icon = self._create_icon("rect", "#696969")
-
-        def refresh_libs():
-            lib_lb.delete(0, tk.END)
-            for lib in self.odd_libraries:
-                lib_lb.insert(tk.END, lib.get("name", ""))
-            refresh_elems()
-
-        def refresh_elems(*_):
-            elem_tree.delete(*elem_tree.get_children())
-            sel = lib_lb.curselection()
-            if not sel:
-                return
-            lib = self.odd_libraries[sel[0]]
-            for el in lib.get("elements", []):
-                name = el.get("name") or el.get("element") or el.get("id")
-                cls = el.get("class", "")
-                attrs = ", ".join(
-                    f"{k}={v}" for k, v in el.items() if k not in {"name", "class"}
-                )
-                opts = {"values": (cls, attrs), "text": name}
-                if self.odd_elem_icon:
-                    opts["image"] = self.odd_elem_icon
-                elem_tree.insert("", tk.END, **opts)
-
-        class ElementDialog(simpledialog.Dialog):
-            def __init__(self, parent, app, data=None):
-                self.app = app
-                self.data = data or {"name": "", "class": ""}
-                super().__init__(parent, title="Edit Element")
-
-            def add_attr_row(self, key="", val=""):
-                r = len(self.attr_rows)
-                k_var = tk.StringVar(value=key)
-                v_var = tk.StringVar(value=str(val))
-                k_entry = ttk.Entry(self.attr_frame, textvariable=k_var)
-                v_entry = ttk.Entry(self.attr_frame, textvariable=v_var)
-                k_entry.grid(row=r, column=0, padx=2, pady=2)
-                v_entry.grid(row=r, column=1, padx=2, pady=2)
-
-                row = {}
-
-                def remove_row():
-                    k_entry.destroy()
-                    v_entry.destroy()
-                    del_btn.destroy()
-                    self.attr_rows.remove(row)
-                    for i, rdata in enumerate(self.attr_rows):
-                        rdata["k_entry"].grid_configure(row=i)
-                        rdata["v_entry"].grid_configure(row=i)
-                        rdata["del_btn"].grid_configure(row=i)
-
-                del_btn = ttk.Button(self.attr_frame, text="Delete", command=remove_row)
-                del_btn.grid(row=r, column=2, padx=2, pady=2)
-
-                row.update(
-                    {
-                        "k_var": k_var,
-                        "v_var": v_var,
-                        "k_entry": k_entry,
-                        "v_entry": v_entry,
-                        "del_btn": del_btn,
-                    }
-                )
-                self.attr_rows.append(row)
-
-            def body(self, master):
-                ttk.Label(master, text="Name").grid(row=0, column=0, sticky="e")
-                self.name_var = tk.StringVar(value=self.data.get("name", ""))
-                ttk.Entry(master, textvariable=self.name_var).grid(row=0, column=1, sticky="ew")
-
-                ttk.Label(master, text="Class").grid(row=1, column=0, sticky="e")
-                self.class_var = tk.StringVar(value=self.data.get("class", ""))
-                cls_opts = [
-                    "Road",
-                    "Infrastructure",
-                    "Temporal",
-                    "Movable",
-                    "Environment",
-                ]
-                ttk.Combobox(master, textvariable=self.class_var, values=cls_opts, state="readonly").grid(row=1, column=1, sticky="ew")
-
-                nb = ttk.Notebook(master)
-                nb.grid(row=2, column=0, columnspan=2, sticky="nsew")
-                master.grid_rowconfigure(2, weight=1)
-                master.grid_columnconfigure(1, weight=1)
-
-                # Attributes tab
-                self.attr_frame = ttk.Frame(nb)
-                nb.add(self.attr_frame, text="Attributes")
-                self.attr_rows = []
-                for k, v in self.data.items():
-                    if k not in {"name", "class", "p", "n", "tp", "fp", "tn", "fn"}:
-                        self.add_attr_row(k, v)
-                ttk.Button(self.attr_frame, text="Add Attribute", command=self.add_attr_row).grid(row=99, column=0, columnspan=3, pady=5)
-
-                # Confusion matrix tab
-                cm_frame = ttk.Frame(nb)
-                nb.add(cm_frame, text="Confusion Matrix")
-                self.p_var = tk.DoubleVar(value=float(self.data.get("p", 0) or 0))
-                self.n_var = tk.DoubleVar(value=float(self.data.get("n", 0) or 0))
-                self.tp_var = tk.DoubleVar(value=float(self.data.get("tp", 0) or 0))
-                self.fp_var = tk.DoubleVar(value=float(self.data.get("fp", 0) or 0))
-                self.tn_var = tk.DoubleVar(value=float(self.data.get("tn", 0) or 0))
-                self.fn_var = tk.DoubleVar(value=float(self.data.get("fn", 0) or 0))
-
-                matrix_metrics = ttk.Frame(cm_frame)
-                matrix_metrics.grid(row=0, column=0, pady=5, sticky="w")
-                matrix = ttk.Frame(matrix_metrics)
-                matrix.grid(row=0, column=0, sticky="w")
-                ttk.Label(matrix, text="").grid(row=0, column=0)
-                ttk.Label(matrix, text="Pred P").grid(row=0, column=1)
-                ttk.Label(matrix, text="Pred N").grid(row=0, column=2)
-                ttk.Label(matrix, text="Actual P").grid(row=1, column=0)
-                ttk.Entry(matrix, textvariable=self.tp_var, width=6).grid(row=1, column=1)
-                ttk.Entry(matrix, textvariable=self.fn_var, width=6).grid(row=1, column=2)
-                ttk.Label(matrix, text="Actual N").grid(row=2, column=0)
-                ttk.Entry(matrix, textvariable=self.fp_var, width=6).grid(row=2, column=1)
-                ttk.Entry(matrix, textvariable=self.tn_var, width=6).grid(row=2, column=2)
-
-                metrics_frame = ttk.Frame(matrix_metrics)
-                metrics_frame.grid(row=0, column=1, padx=10, sticky="n")
-                ttk.Label(metrics_frame, text="Accuracy:").grid(row=0, column=0, sticky="e")
-                ttk.Label(metrics_frame, text="Precision:").grid(row=1, column=0, sticky="e")
-                ttk.Label(metrics_frame, text="Recall:").grid(row=2, column=0, sticky="e")
-                ttk.Label(metrics_frame, text="F1 Score:").grid(row=3, column=0, sticky="e")
-                self.acc_var = tk.StringVar()
-                self.prec_var = tk.StringVar()
-                self.rec_var = tk.StringVar()
-                self.f1_var = tk.StringVar()
-                ttk.Label(metrics_frame, textvariable=self.acc_var).grid(row=0, column=1, sticky="w")
-                ttk.Label(metrics_frame, textvariable=self.prec_var).grid(row=1, column=1, sticky="w")
-                ttk.Label(metrics_frame, textvariable=self.rec_var).grid(row=2, column=1, sticky="w")
-                ttk.Label(metrics_frame, textvariable=self.f1_var).grid(row=3, column=1, sticky="w")
-
-                def update_metrics(*_):
-                    from analysis.confusion_matrix import compute_metrics
-
-                    tp = self.tp_var.get()
-                    fp = self.fp_var.get()
-                    tn = self.tn_var.get()
-                    fn = self.fn_var.get()
-                    metrics = compute_metrics(tp, fp, tn, fn)
-                    self.acc_var.set(f"{metrics['accuracy']:.3f}")
-                    self.prec_var.set(f"{metrics['precision']:.3f}")
-                    self.rec_var.set(f"{metrics['recall']:.3f}")
-                    self.f1_var.set(f"{metrics['f1']:.3f}")
-                    self.p_var.set(tp + fn)
-                    self.n_var.set(tn + fp)
-
-                for var in (self.tp_var, self.fp_var, self.tn_var, self.fn_var):
-                    var.trace_add("write", update_metrics)
-                update_metrics()
-
-                vt_frame = ttk.Frame(cm_frame)
-                vt_frame.grid(row=1, column=0, sticky="nsew", pady=5)
-                cm_frame.grid_rowconfigure(1, weight=1)
-                cm_frame.grid_columnconfigure(0, weight=1)
-                columns = [
-                    "Product Goal",
-                    "Validation Target",
-                    "Target Description",
-                    "Acceptance Criteria",
-                ]
-                self.vt_tree = ttk.Treeview(
-                    vt_frame, columns=columns, show="headings", height=4
-                )
-                for c in columns:
-                    self.vt_tree.heading(c, text=c)
-                    width = 120 if c in ("Product Goal", "Validation Target") else 200
-                    self.vt_tree.column(c, width=width, anchor="center")
-                self.vt_tree.pack(fill=tk.BOTH, expand=True)
-                self.vt_item_to_goal = {}
-                self.selected_goal = None
-                self.current_tau_on = 0.0
-
-                def on_vt_select(event=None):
-                    sel = self.vt_tree.selection()
-                    if not sel:
-                        self.selected_goal = None
-                        self.current_tau_on = 0.0
-                    else:
-                        sg = self.vt_item_to_goal.get(sel[0])
-                        self.selected_goal = sg
-                        tau_on = 0.0
-                        mp_name = getattr(sg, "mission_profile", "")
-                        if mp_name:
-                            for mp in self.app.mission_profiles:
-                                if mp.name == mp_name:
-                                    tau_on = mp.tau_on
-                                    break
-                        self.current_tau_on = tau_on
-                    update_metrics()
-
-                self.vt_tree.bind("<<TreeviewSelect>>", on_vt_select)
-
-                def refresh_vt(*_):
-                    self.vt_tree.delete(*self.vt_tree.get_children())
-                    self.vt_item_to_goal.clear()
-                    name = self.name_var.get().strip()
-                    for sg in self.app.get_validation_targets_for_odd(name):
-                        iid = self.vt_tree.insert(
-                            "",
-                            "end",
-                            values=[
-                                sg.user_name or f"SG {sg.unique_id}",
-                                getattr(sg, "validation_target", ""),
-                                getattr(sg, "validation_desc", ""),
-                                getattr(sg, "acceptance_criteria", ""),
-                            ],
-                        )
-                        self.vt_item_to_goal[iid] = sg
-                    items = self.vt_tree.get_children()
-                    if items:
-                        self.vt_tree.selection_set(items[0])
-                        on_vt_select()
-
-                refresh_vt()
-                self.name_var.trace_add("write", refresh_vt)
-
-            def apply(self):
-                new_data = {"name": self.name_var.get(), "class": self.class_var.get()}
-                for row in self.attr_rows:
-                    key = row["k_var"].get().strip()
-                    if key:
-                        new_data[key] = row["v_var"].get()
-                tp = float(self.tp_var.get())
-                fp = float(self.fp_var.get())
-                tn = float(self.tn_var.get())
-                fn = float(self.fn_var.get())
-                from analysis.confusion_matrix import compute_metrics
-
-                metrics = compute_metrics(tp, fp, tn, fn)
-                p = tp + fn
-                n = tn + fp
-                new_data.update({
-                    "tp": tp,
-                    "fp": fp,
-                    "tn": tn,
-                    "fn": fn,
-                    "p": p,
-                    "n": n,
-                })
-                new_data.update(metrics)
-                self.data = new_data
-
-        def add_lib():
-            name = simpledialog.askstring("New Library", "Library name:")
-            if not name:
-                return
-            elems = []
-            if messagebox.askyesno("Import", "Import elements from file?"):
-                path = filedialog.askopenfilename(filetypes=[("CSV/Excel", "*.csv *.xlsx")])
-                if path:
-                    if path.lower().endswith(".csv"):
-                        with open(path, newline="") as f:
-                            elems = list(csv.DictReader(f))
-                    elif path.lower().endswith(".xlsx"):
-                        try:
-                            if load_workbook is None:
-                                raise ImportError
-                            wb = load_workbook(path, read_only=True)
-                            ws = wb.active
-                            headers = [c.value for c in next(ws.iter_rows(max_row=1))]
-                            for row in ws.iter_rows(min_row=2, values_only=True):
-                                elem = {headers[i]: row[i] for i in range(len(headers))}
-                                elems.append(elem)
-                        except Exception:
-                            messagebox.showerror("Import", "Failed to read Excel file. openpyxl required.")
-            self.odd_libraries.append({"name": name, "elements": elems})
-            refresh_libs()
-            self.update_odd_elements()
-
-        def edit_lib():
-            sel = lib_lb.curselection()
-            if not sel:
-                return
-            lib = self.odd_libraries[sel[0]]
-            name = simpledialog.askstring("Edit Library", "Library name:", initialvalue=lib.get("name", ""))
-            if name:
-                lib["name"] = name
-                refresh_libs()
-
-        def delete_lib():
-            sel = lib_lb.curselection()
-            if sel:
-                idx = sel[0]
-                del self.odd_libraries[idx]
-                refresh_libs()
-                self.update_odd_elements()
-
-        def add_elem():
-            sel = lib_lb.curselection()
-            if not sel:
-                return
-            lib = self.odd_libraries[sel[0]]
-            dlg = ElementDialog(win, self)
-            lib.setdefault("elements", []).append(dlg.data)
-            refresh_elems()
-            self.update_odd_elements()
-
-        def edit_elem():
-            sel_lib = lib_lb.curselection()
-            sel_elem = elem_tree.selection()
-            if not sel_lib or not sel_elem:
-                return
-            lib = self.odd_libraries[sel_lib[0]]
-            idx = elem_tree.index(sel_elem[0])
-            data = lib.get("elements", [])[idx]
-            dlg = ElementDialog(win, self, data)
-            lib["elements"][idx] = dlg.data
-            refresh_elems()
-            self.update_odd_elements()
-
-        def del_elem():
-            sel_lib = lib_lb.curselection()
-            sel_elem = elem_tree.selection()
-            if not sel_lib or not sel_elem:
-                return
-            lib = self.odd_libraries[sel_lib[0]]
-            idx = elem_tree.index(sel_elem[0])
-            del lib.get("elements", [])[idx]
-            refresh_elems()
-            self.update_odd_elements()
-
-        btnf = ttk.Frame(win)
-        btnf.grid(row=1, column=1, columnspan=3, sticky="ew")
-        ttk.Button(btnf, text="Add Lib", command=add_lib).pack(side=tk.LEFT)
-        ttk.Button(btnf, text="Edit Lib", command=edit_lib).pack(side=tk.LEFT)
-        ttk.Button(btnf, text="Del Lib", command=delete_lib).pack(side=tk.LEFT)
-        ttk.Button(btnf, text="Add Elem", command=add_elem).pack(side=tk.LEFT, padx=5)
-        ttk.Button(btnf, text="Edit Elem", command=edit_elem).pack(side=tk.LEFT)
-        ttk.Button(btnf, text="Del Elem", command=del_elem).pack(side=tk.LEFT)
-
-        lib_lb.bind("<<ListboxSelect>>", refresh_elems)
-        refresh_libs()
+        return self.odd_library_manager.manage_odd_libraries()
 
     def open_reliability_window(self):
         return self.open_windows_features.open_reliability_window()
