+++ mainappsrc/version.py
@@ -18,6 +18,6 @@
 
 """Project version information."""
 
-VERSION = "0.2.54"
+VERSION = "0.2.55"
 
 __all__ = ["VERSION"]
